GregoryLoredo

Documentation for GregoryLoredo.

GregoryLoredo

GregoryLoredo.lcshape

GregoryLoredo.orsum

GregoryLoredo.persum

GregoryLoredo.BinMult

GregoryLoredo.BinningFactor

GregoryLoredo.ComputeBinValues

GregoryLoredo.FrequencyRange

GregoryLoredo.FullBinMult

GregoryLoredo.FullBinMultLog

GregoryLoredo.LightCurveShape

GregoryLoredo.LightCurveSimpleShape

GregoryLoredo.MarginalizedPeriodogram

GregoryLoredo.OddRatioSummary

GregoryLoredo.OddRatios

GregoryLoredo.Periodogram

GregoryLoredo.PeriodogramSummary

GregoryLoredo.PlotLightCurve

GregoryLoredo.PlotOddRatios

GregoryLoredo.PlotPeriodogram

GregoryLoredo.WeightingFactor

GregoryLoredo.jt

GregoryLoredo.logStirlingApprox

GregoryLoredo.lcshape — Type

lcshape

Fields

μlc: Vector of rate mean value for each bin

σlc: Vector of standard deviation for the rate for each bin

GregoryLoredo.orsum — Type

orsum

Fields

prob: Vector of probabilities for each possible bin number m

maxm: Bin number showing the maximum probability

GregoryLoredo.persum — Type

persum

Fields

maxpow: Maximum power in the periodogram

maxfreq: Frequency at the maximum power

maxper: Period at the maximum power

GregoryLoredo.BinMult — Method

BinMult(nbins::Vector{Int})

#Arguments

nbins is a vector with the number of events for each bin.

Compute the basic factor for the multiplicity of the binned events (Eq. 1.4 in Gregory & Lorendo (1992)).

GregoryLoredo.BinningFactor — Method

BinningFactor(ω::Float64,Φ::Float64,m::Int,tlist::Vector{Float64})::BigFloat

#Arguments

ω is the angular frequency.

Φ is the phase.

m is the number of bins.

tlist is the vector of arrival times.

Compute the binning factor (Eq. B4 in Gregory & Lorendo (1992)).

GregoryLoredo.ComputeBinValues — Method

ComputeBinValues(m::Int,ω::Float64,Φ::Float64,tlist::Vector{Float64})

#Arguments

m is the number of bins.

ω is the angular frequency.

Φ is the phase.

tlist is the vector of arrival times.

Compute the number of events falling in each bin of a model with m bins with given frequency and phase.

GregoryLoredo.FrequencyRange — Method

FrequencyRange(Tlist::Vector{Float64};tstart=minimum(Tlist),tend=maximum(Tlist),logspacing=false))

#Arguments

Tlist is the vector of arrival times.

tstart is the minimum possible arrival time (default to the minimum of Tlist)

tend is the minimum possible arrival time (default to the maximum of Tlist)

logspacing generates a log-spaced frquency range.

Compute the frequency range to be analysed accordiong to the advises in Gregory & Lorendo (1992)).

GregoryLoredo.FullBinMult — Method

FullBinMult(nbins::Vector{Int})

#Arguments

nbins is a vector with the number of events for each bin.

Compute the full multiplicity of the binned events (Eq. 1.4 in Gregory & Lorendo (1992)).

GregoryLoredo.FullBinMultLog — Method

FullBinMultLog(nbins::Vector{Int})

#Arguments

nbins is a vector with the number of events for each bin.

Compute the full multiplicity of the binned events (Eq. 1.4 in Gregory & Lorendo (1992)).

The computation is here carried out by the logarithm.

GregoryLoredo.LightCurveShape — Method

LightCurveShape(ωr::Vector{Float64},m::Int,tlist::Vector{Float64};unc=false,progress=false,phasesteps=10)::lcshape

#Arguments

ωr is the vector of the analyzed frequencies.

m is the number of bins.

tlist is the vector of arrival times.

unc computes light-curve uncertainty if selected.

progress shows progress bars if selected.

Compute the light curve shape given the number m of bins and the input arrival time tlist (Eq. 7.10 in Gregory & Lorendo (1992)).

GregoryLoredo.LightCurveSimpleShape — Method

LightCurveSimpleShape(ω::Float64,Φ::Float64,m::Int,tlist::Vector{Float64};unc=false)::lcshape

#Arguments

ω is the angular frequency.

Φ is the phase.

m is the number of bins.

tlist is the vector of arrival times.

unc computes light-curve uncertainty if selected.

Compute the light curve shape given the angular frequency ω, the phase Φ, m bins given the input arrival time tlist (Eq. 4.3 in Gregory & Lorendo (1992)).

GregoryLoredo.MarginalizedPeriodogram — Method

MarginalizedPeriodogram(Tlist::Vector{Float64},ωr::Vector{Float64};progress=false,phasesteps=10)

#Arguments

Tlist is the vector of arrival times.

ωr is the vector of the analyzed frequencies.

mmax is the maximum number of bins.

progress shows progress bars if selected.

Compute the periodogram following Section 6 in Gregory & Lorendo (1992)).

GregoryLoredo.OddRatioSummary — Method

OddRatioSummary(or::Vector{Real})::orsum

#Arguments

od is the vector of computed odd ratios for each considered light-curve bin number.

Compute the probability for a non constant model Eq. 5.2. in Gregory & Lorendo (1992)) and the bin number yileding the highest odd-ratio.

GregoryLoredo.OddRatios — Method

OddRatios(Tlist::Vector{Float64},ωr::Vector{Float64},mmax::Int;progress=false,phasesteps=10)

#Arguments

Tlist is the vector of arrival times.

ωr is the vector of the analyzed frequencies.

mmax is the maximum number of bins.

progress shows progress bars if selected.

Compute the odd ratio following Appendix A in Gregory & Lorendo (1992)).

GregoryLoredo.Periodogram — Function

Periodogram(Tlist::Vector{Float64},ωr::Vector{Float64},m::Int,phasesteps=10)

#Arguments

Tlist is the vector of arrival times.

ωr is the vector of the analyzed frequencies.

m is the number of bins.

Compute the periodogram following Section 6 in Gregory & Lorendo (1992)).

GregoryLoredo.PeriodogramSummary — Method

PeriodogramSummary(per::Vector{BigFloat},ωr::Vector{Float64})::persum

#Arguments

per is the vector of powers for each considered frequency.

ωr is the vector of the analyzed frequencies.

Plot the Odd Ratio as a function of the number of bins.

GregoryLoredo.PlotLightCurve — Method

PlotLightCurve(lpl::lcshape;xlabel="",ylabel="",sigma=1.)

#Arguments

lpl is a light-curve shape structure.

xlabel is the (optional) label for the x-axis.

ylabel is the (optional) label for the y-axis.

sigma is the number od standard deviation to be plotted.

Plot the light-curve in the given bins of the analysis.

GregoryLoredo.PlotOddRatios — Method

PlotOddRatios(or::Vector{Real};xlabel="",ylabel="";xlabel="",ylabel="")

#Arguments

od is the vector of computed odd ratios for each considered light-curve bin number.

xlabel is the (optional) label for the x-axis.

ylabel is the (optional) label for the y-axis.

Plot the oddratio for each possible bin number.

GregoryLoredo.PlotPeriodogram — Method

PlotPeriodogram(per::Vector{BigFloat},ωr::Vector{Float64};xlabel="",ylabel="",pmax=2π/minimum(ωr),pmin=2π/maximum(ωr))

#Arguments

per is the vector of powers for each considered frequency.

ωr is the vector of the analyzed frequencies.

xlabel is the (optional) label for the x-axis.

ylabel is the (optional) label for the y-axis.

pmin is the minimum of the plotted periods (default minimum of the input data).

pmax is the maximum of the plotted periods (default maximum of the input data).

Plot the periodogram as a function of the analysed frequencies.

GregoryLoredo.WeightingFactor — Method

WeightingFactor(njbin::Vector{Int},m::Int,totcnt::Int)::Vector{BigFloat}

#Arguments

njbin is the vector with the number of events in each model bin.

m is the number of bins for the adopted model.

totcnt is the total number of events in the analyzed dataset.

Compute the weighting factors for each bin of the model.

GregoryLoredo.jt — Method

jt(m::Int,ω::Float64,Φ::Float64,t::Float64)

#Arguments

m is the number of bins.

ω is the angular frequency.

Φ is the phase.

t is the time.

Compute the bin number at time tfor a model with m bins with given frequency and phase.

GregoryLoredo.logStirlingApprox — Method

logStirlingApprox(n::Int)

#Arguments

n is the integer we want to use to compute the factorial.